+++ code/test/sixsq/nuvla/server/resources/user_template_github_lifecycle_test.clj
@@ -58,7 +58,7 @@
                                  session
                                  (content-type "application/json")
                                  (header authn-info/authn-info-header "unknown group/nuvla-anon"))
-        session-admin        (header session-anon authn-info/authn-info-header "super group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-admin        (header session-anon authn-info/authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
         session-user         (header session-anon authn-info/authn-info-header "user group/nuvla-user group/nuvla-anon")
 
         redirect-url-example "https://example.com/webui"]